requirements.txt
nltk==3.8.1
numpy==1.26.4
tensorflow-cpu==2.16.1
flask==3.0.0
setuptools>=69.0.0
